HR System/HR System Code/requirements.txt
blis==0.7.4
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.4
ci-info==0.2.0
click==7.1.2
cobble==0.1.3
colorama==0.4.4
configobj==5.0.6
configparser==5.0.2
cymem==2.0.5
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
Flask==2.0.1
future==0.18.2
fuzzywuzzy==0.18.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.8.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.3
mammoth==1.4.17
MarkupSafe==2.0.1
murmurhash==1.0.5
networkx==2.6.2
nibabel==3.2.1
nipype==1.6.1
nltk==3.6.2
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pathlib==1.0.1
pathy==0.6.0
preshed==3.0.5
prov==2.0.0
pydantic==1.8.2
pydot==1.4.2
PyMuPDF==1.18.17
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pyxnat==1.4
rdflib==6.0.0
regex==2021.8.28
requests==2.26.0
scipy==1.7.1
simplejson==3.17.5
six==1.16.0
smart-open==5.2.1
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.8
tqdm==4.62.2
traits==6.2.0
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
zipp==3.5.0
